Roles and permissions: added before excel reports

+++ app/routes/_app+/agents/$agentRouteId/_index.tsx
@@ -3,6 +3,7 @@ import { redirectWithWarning } from "remix-toast";
 import {  handlerErrorWithToast } from "~/.server/reponses/handlerError";
 import { handlerSuccessWithToast } from "~/.server/reponses/handlerSuccess";
 import { Service } from "~/.server/services";
+import { permissions } from "~/application";
 
 
 export const action: ActionFunction = async({params, request}) => {
@@ -13,6 +14,7 @@ export const action: ActionFunction = async({params, request}) => {
     try {
   
       if(data._action === 'delete') {
+        await Service.auth.requirePermission(request, permissions.agents.permissions.delete);
         await Service.agent.deleteOne(id);
         return handlerSuccessWithToast('delete');
       }
