+++ app/routes/_app+/clients_.$creditId/_index.tsx
@@ -9,13 +9,17 @@ import {
   handlerErrorWithToast 
 } from "~/.server/reponses";
 import { Service } from "~/.server/services";
+import { permissions } from "~/application";
 import { ChipStatusCredit, PersonFormEdit } from "~/components/ui";
 import { CreditFormEdit } from "~/components/ui/credit/CreditFormEdit";
 import { ButtonAddPayment, ModalPay } from "~/components/ui/pay";
+import { Permission } from '~/components/ui/auth/Permission';
+import { ErrorBoundary } from './payments/_layout';
 
 
-export const loader: LoaderFunction = async ({ params }) => {
+export const loader: LoaderFunction = async ({ request, params }) => {
   
+  await Service.auth.requirePermission(request, permissions.credits.permissions.view_detail);
   const { creditId } = params;
 
   try {
@@ -38,11 +42,13 @@ export const action: ActionFunction = async ({ params, request }) => {
   try {
     
       if(data._action === 'delete') {
+        await Service.auth.requirePermission(request, permissions.credits.permissions.delete);
         await Service.credit.deleteOne(id);
         return handlerSuccessWithToast('delete');
       }
       
       if(data._action === 'update') {
+        await Service.auth.requirePermission(request, permissions.credits.permissions.update);
         await Service.credit.updateOne(formData, id);
         return handlerSuccessWithToast('update');
       }
@@ -56,7 +62,7 @@ export const action: ActionFunction = async ({ params, request }) => {
 
 }
 
-
+export { ErrorBoundary };
 export default function CreditDetailPage() {
   const loader = useLoaderData<any>();
 
@@ -77,7 +83,9 @@ export default function CreditDetailPage() {
             />
             {
               (credit.currentDebt > 0) && (
-                <ButtonAddPayment creditId={credit.id} onPress={onOpen}  />
+                <Permission permission={permissions.payments.permissions.add}>
+                  <ButtonAddPayment creditId={credit.id} onPress={onOpen}  />
+                </Permission>
               ) 
             }
           </div>
@@ -90,23 +98,24 @@ export default function CreditDetailPage() {
       <CardBody className="flex flex-col gap-2">
         <ModalPay isOpen={isOpen} onOpenChange={onOpenChange} />
           <PersonFormEdit 
-            { ...client }
-            urlAction="clients"
-            title='Datos del cliente'
-            />
+              { ...client }
+              urlAction="clients"
+              title='Datos del cliente'
+              permission={permissions.credits.permissions.update_client}
+          />
           <PersonFormEdit 
-            { ...aval }
-            urlAction="aval"
-            title='Datos del aval'
+              { ...aval }
+              urlAction="aval"
+              title='Datos del aval'
+              permission={permissions.credits.permissions.update_aval}
           />
         <CreditFormEdit 
           { ...credit }
         />
       </CardBody>
-      <CardFooter>
-          Visit source code on GitHub.
-      </CardFooter>
     </Card>
-    <Outlet context={{ client: client.id, folder: credit.folder.id, group:  credit.group.id }}/>
+    <Permission permission={permissions.payments.permissions.view}>
+      <Outlet context={{ client: client.id, folder: credit.folder.id, group:  credit.group.id }}/>
+    </Permission>
   </>)
 }
